+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.1.2] - 2025-06-28
+### Fixed
+- Better handling of bytes types in packing calculation
+- Added options --list-rules
+
 ## [0.1.1] - 2025-06-21
 ### Fixed
 - Added missing rules to rule list